Module polkadot_parachain::chain_spec

Structs

Extensions

The extensions for the ChainSpec.

ExtensionsFork

Constants

CANVAS_KUSAMA_ED

The existential deposit is determined by the runtime “canvas-kusama”.

CANVAS_PARACHAIN_ID

No relay chain suffix because the id is the same over all relay chains.

SAFE_XCM_VERSION

The default XCM version to set in genesis config.

STATEMINE_ED

STATEMINT_ED

WESTMINT_ED

Functions

canvas_kusama_genesis

canvas_rococo_config

canvas_rococo_development_config

canvas_rococo_local_config

get_account_id_from_seed

Helper function to generate an account ID from seed

get_chain_spec

get_collator_keys_from_seed

Generate collator keys from seed.

get_from_seed

Helper function to generate a crypto pair from seed

get_public_from_seed

Helper function to generate a crypto pair from seed

get_seedling_chain_spec

get_shell_chain_spec

seedling_testnet_genesis

shell_testnet_genesis

staging_test_net

statemine_config

statemine_development_config

statemine_genesis

statemine_local_config

statemine_session_keys

Generate the session keys from individual elements.

statemint_config

statemint_development_config

statemint_genesis

statemint_local_config

statemint_session_keys

Generate the session keys from individual elements.

testnet_genesis

westmint_config

westmint_development_config

westmint_genesis

westmint_local_config

westmint_session_keys

Generate the session keys from individual elements.

Type Definitions

AccountPublic

CanvasKusamaChainSpec

We use the same runtime on kusama and rococo.

ChainSpec

Specialized ChainSpec for the normal parachain runtime.

SeedlingChainSpec

Specialized ChainSpec for the seedling parachain runtime.

ShellChainSpec

Specialized ChainSpec for the shell parachain runtime.

StatemineChainSpec

StatemintChainSpec

Specialized ChainSpec for the normal parachain runtime.

WestmintChainSpec