Function polkadot_parachain::service::build_relay_to_aura_import_queue

pub fn build_relay_to_aura_import_queue<RuntimeApi, AuraId: AppCrypto>(
    client: Arc<TFullClient<Block, RuntimeApi, WasmExecutor<(SubstrateHostFunctions, HostFunctions)>>>,
    block_import: ParachainBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<(SubstrateHostFunctions, HostFunctions)>>>, TFullBackend<Block>>,
    config: &Configuration,
    telemetry_handle: Option<TelemetryHandle>,
    task_manager: &TaskManager
) -> Result<DefaultImportQueue<Block>, Error>where
    RuntimeApi: ConstructRuntimeApi<Block, TFullClient<Block, RuntimeApi, WasmExecutor<(SubstrateHostFunctions, HostFunctions)>>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: TaggedTransactionQueue<Block> + Metadata<Block> + SessionKeys<Block> + ApiExt<Block> + OffchainWorkerApi<Block> + BlockBuilder<Block> + AuraApi<Block, <<AuraId as AppCrypto>::Pair as Pair>::Public>,
    <<AuraId as AppCrypto>::Pair as Pair>::Signature: TryFrom<Vec<u8>> + Hash + Member + Codec,

Build the import queue for parachain runtimes that started with relay chain consensus and switched to aura.