Function polkadot_parachain::service::start_node_impl

async fn start_node_impl<RuntimeApi, RB, BIQ, SC, Net>(
    parachain_config: Configuration,
    polkadot_config: Configuration,
    collator_options: CollatorOptions,
    sybil_resistance_level: CollatorSybilResistance,
    para_id: ParaId,
    rpc_ext_builder: RB,
    build_import_queue: BIQ,
    start_consensus: SC,
    hwbench: Option<HwBench>
) -> Result<(TaskManager, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>)>
where
    RuntimeApi: ConstructRuntimeApi<Block, TFullClient<Block, RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: TaggedTransactionQueue<Block> + Metadata<Block> + SessionKeys<Block> + ApiExt<Block> + OffchainWorkerApi<Block> + BlockBuilder<Block> + CollectCollationInfo<Block> + TransactionPaymentRuntimeApi<Block, Balance> + AccountNonceApi<Block, AccountId, Nonce>,
    RB: Fn(DenyUnsafe, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>, Arc<TFullBackend<Block>>, Arc<FullPool<Block, TFullClient<Block, RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>>) -> Result<RpcModule<()>, Error> + 'static,
    BIQ: FnOnce(Arc<TFullClient<Block, RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>, ParachainBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>, TFullBackend<Block>>, &Configuration, Option<TelemetryHandle>, &TaskManager) -> Result<DefaultImportQueue<Block>, Error>,
    SC: FnOnce(Arc<TFullClient<Block, RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>, ParachainBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>, TFullBackend<Block>>, Option<&Registry>, Option<TelemetryHandle>, &TaskManager, Arc<dyn RelayChainInterface>, Arc<FullPool<Block, TFullClient<Block, RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>>, Arc<SyncingService<Block>>, KeystorePtr, Duration, ParaId, CollatorPair, OverseerHandle, Arc<dyn Fn(Hash, Option<Vec<u8>>) + Send + Sync>, Arc<TFullBackend<Block>>) -> Result<(), Error>,
    Net: NetworkBackend<Block, Hash>,

Start a node with the given parachain Configuration and relay chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api.