Function polkadot_parachain::service::start_shell_node_impl

async fn start_shell_node_impl<RuntimeApi, RB, BIQ, SC>(
    parachain_config: Configuration,
    polkadot_config: Configuration,
    collator_options: CollatorOptions,
    sybil_resistance_level: CollatorSybilResistance,
    para_id: ParaId,
    rpc_ext_builder: RB,
    build_import_queue: BIQ,
    start_consensus: SC,
    hwbench: Option<HwBench>
) -> Result<(TaskManager, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>)>
where
    RuntimeApi: ConstructRuntimeApi<Block, TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: TaggedTransactionQueue<Block> + Metadata<Block> + SessionKeys<Block> + ApiExt<Block> + OffchainWorkerApi<Block> + BlockBuilder<Block> + CollectCollationInfo<Block>,
    RB: Fn(Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>) -> Result<RpcModule<()>, Error> + 'static,
    BIQ: FnOnce(Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, ParachainBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, TFullBackend<Block>>, &Configuration, Option<TelemetryHandle>, &TaskManager) -> Result<DefaultImportQueue<Block>, Error>,
    SC: FnOnce(Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, ParachainBlockImport<Block, Arc<TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>, TFullBackend<Block>>, Option<&Registry>, Option<TelemetryHandle>, &TaskManager, Arc<dyn RelayChainInterface>, Arc<FullPool<Block, TFullClient<Block, RuntimeApi, WasmExecutor<SubstrateHostFunctions>>>>, Arc<SyncingService<Block>>, KeystorePtr, Duration, ParaId, CollatorPair, OverseerHandle, Arc<dyn Fn(Hash, Option<Vec<u8>>) + Send + Sync>) -> Result<(), Error>,

Start a shell node with the given parachain Configuration and relay chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api for shell nodes.