Trait polkadot_parachain::service::NodeSpec

pub(crate) trait NodeSpec {
    type RuntimeApi: ConstructNodeRuntimeApi<Block, TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>;
    type BuildImportQueue: BuildImportQueue<Self::RuntimeApi> + 'static;
    type BuildRpcExtensions: BuildRpcExtensions<TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>, TFullBackend<Block>, FullPool<Block, TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>> + 'static;
    type StartConsensus: StartConsensus<Self::RuntimeApi> + 'static;

    const SYBIL_RESISTANCE: CollatorSybilResistance;

    // Provided methods
    fn new_partial(
        config: &Configuration
    ) -> Result<PartialComponents<TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>, TFullBackend<Block>, (), DefaultImportQueue<Block>, FullPool<Block, TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>, (ParachainBlockImport<Block, Arc<TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>, TFullBackend<Block>>, Option<Telemetry>, Option<TelemetryWorkerHandle>)>> { ... }
    fn start_node<Net>(
        parachain_config: Configuration,
        polkadot_config: Configuration,
        collator_options: CollatorOptions,
        para_id: ParaId,
        hwbench: Option<HwBench>
    ) -> Pin<Box<dyn Future<Output = Result<TaskManager>>>>
       where Net: NetworkBackend<Block, Hash> { ... }
}

Required Associated Types

type RuntimeApi: ConstructNodeRuntimeApi<Block, TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>

type BuildImportQueue: BuildImportQueue<Self::RuntimeApi> + 'static

type BuildRpcExtensions: BuildRpcExtensions<TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>, TFullBackend<Block>, FullPool<Block, TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>> + 'static

type StartConsensus: StartConsensus<Self::RuntimeApi> + 'static

Required Associated Constants

const SYBIL_RESISTANCE: CollatorSybilResistance

Provided Methods

fn new_partial( config: &Configuration ) -> Result<PartialComponents<TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>, TFullBackend<Block>, (), DefaultImportQueue<Block>, FullPool<Block, TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>, (ParachainBlockImport<Block, Arc<TFullClient<Block, Self::RuntimeApi, WasmExecutor<(ParachainHostFunctions, HostFunctions)>>>, TFullBackend<Block>>, Option<Telemetry>, Option<TelemetryWorkerHandle>)>>

Starts a ServiceBuilder for a full service.

Use this macro if you don’t actually need the full service, but just the builder in order to be able to perform chain operations.

fn start_node<Net>( parachain_config: Configuration, polkadot_config: Configuration, collator_options: CollatorOptions, para_id: ParaId, hwbench: Option<HwBench> ) -> Pin<Box<dyn Future<Output = Result<TaskManager>>>>
where Net: NetworkBackend<Block, Hash>,

Start a node with the given parachain spec.

This is the actual implementation that is abstract over the executor and the runtime api.

Object Safety

This trait is not object safe.