Type Alias polkadot_parachain::service::AccountId

pub type AccountId = <<MultiSignature as Verify>::Signer as IdentifyAccount>::AccountId;

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.