Module polkadot_parachain::service

Structs

Verifier 🔒

WaitForAuraConsensus 🔒

Special ParachainConsensus implementation that waits for the upgrade from shell to a parachain runtime that implements Aura.

Enums

BuildOnAccess 🔒

Functions

build_aura_import_queue

Build the import queue for Aura-based runtimes.

build_contracts_rpc_extensions 🔒

build_parachain_rpc_extensions 🔒

build_relay_to_aura_import_queue

Build the import queue for parachain runtimes that started with relay chain consensus and switched to aura.

build_shell_import_queue

Build the import queue for the shell runtime.

new_partial

Starts a ServiceBuilder for a full service.

start_asset_hub_lookahead_node

Start a shell node which should later transition into an Aura powered parachain node. Asset Hub uses this because at genesis, Asset Hub was on the shell runtime which didn’t have Aura and needs to sync and upgrade before it can run AuraApi functions.

start_basic_lookahead_node

Start an aura powered parachain node which uses the lookahead collator to support async backing. This node is basic in the sense that its runtime api doesn’t include common contents such as transaction payment. Used for aura glutton.

start_contracts_rococo_node

Start a parachain node for Rococo Contracts.

start_generic_aura_lookahead_node

Uses the lookahead collator to support async backing.

start_lookahead_aura_consensus 🔒

Start consensus using the lookahead aura collator.

start_node_impl 🔒

Start a node with the given parachain Configuration and relay chain Configuration.

start_relay_chain_consensus 🔒

Start relay-chain consensus that is free for all. Everyone can submit a block, the relay-chain decides what is backed and included.

start_rococo_parachain_node

Start a rococo parachain node.

start_shell_node

Start a polkadot-shell parachain node.

warn_if_slow_hardware 🔒

Checks that the hardware meets the requirements and print a warning otherwise.

Type Aliases

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AuraId

Balance

Balance of an account.

Block

Opaque block type.

Hash

A hash of some data used by the chain.

Header

Opaque block header type.

HostFunctions 🔒

Nonce

Index of a transaction in the chain.

ParachainBackend 🔒

ParachainBlockImport 🔒

ParachainClient 🔒

Service

Assembly of PartialComponents (enough to run chain ops subcommands)